VertexAIEmbeddings unit tests
